Add doc
Add doc

+++ src/main/java/example/micronaut/repository/entity/Counter.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2023 denjossal
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package example.micronaut.repository.entity;
 
 import io.micronaut.core.annotation.NonNull;
@@ -18,9 +34,9 @@ public class Counter {
 
     @NonNull
     @NotBlank
-    private final BigInteger c;
+    private final Integer c;
 
-    public Counter(String id, BigInteger countNumber, BigInteger c) {
+    public Counter(String id, BigInteger countNumber, Integer c) {
         this.id = id;
         this.countNumber = countNumber;
         this.c = c;
@@ -30,11 +46,13 @@ public class Counter {
         return id;
     }
 
-    public BigInteger getCountNumber(){
+    public BigInteger getCountNumber() {
         return countNumber;
     }
 
-    public BigInteger getC() { return c; }
+    public Integer getC() {
+        return c;
+    }
 
     @Override
     public String toString() {
